refactor(web): group inderun-web src by concern (core/providers)
25 flat files made it hard to see engine-core vs host vs per-provider
boundaries. Move engine/router/registry/provider/host/errors/telemetry/
route-planner into core/, and each provider family (openai, onnx,
system-model) into providers/<name>/. Public subpath entrypoints
(index.ts, openai.ts, onnx.ts, system-model.ts, web-sdk.ts) stay at src
root unchanged so dist/* paths in package.json exports keep resolving
the same way. Pure move, only import paths changed.
