+++ packages/inderun-web/src/onnx.ts
@@ -12,7 +12,7 @@ export {
   DEFAULT_ONNX_WEB_PROVIDER_ID,
   SUPPORTED_WEB_MODEL_SOURCE_TYPES,
   type OnnxProviderOptions
-} from "./onnx-provider.js";
+} from "./providers/onnx/provider.js";
 
 export {
   OnnxRuntimeError,
@@ -24,11 +24,11 @@ export {
   type OnnxRuntimeAvailability,
   type OnnxRuntimeErrorKind,
   type OnnxTextGenerationRuntime
-} from "./onnx-runtime.js";
+} from "./providers/onnx/runtime.js";
 
 export {
   createTransformersJsRuntime,
   type TransformersJsRuntimeOptions,
   type TransformersModule,
   type TransformersTextGenerator
-} from "./onnx-transformers-runtime.js";
+} from "./providers/onnx/transformers-runtime.js";
